+++ cli/lib/ask.js
@@ -1,6 +1,6 @@
 const async = require('async')
 const inquirer = require('inquirer')
-const evaluate = require('./eval')
+const evaluate = require('./util/eval')
 
 // Support types from prompt-for which was used before
 const promptMapping = {
