--- cli/lib/src/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "{{name}}",
-  "description": "{{description}}",
-  "repository": "git://github.com/{{repository}}.git",
-  "version": "0.0.0",
-  "license": "{{license}}",
-  "main": "lib/index.js",
-  "dependencies": {},
-  "devDependencies": {
-    "mocha": "1.x"
-  }
-}
