+++ cli/lib/template.js
@@ -3,10 +3,6 @@ const { fetchRepoList } = require('./util/gitTemplate')
 const { clearConsole } = require('./util/logger')
 const { logWithSpinner,stopSpinner } = require('./util/spinner')
 module.exports = async ()=>{
-
-  // process.on('exit', () => {
-  //   console.log()
-  // })
   logWithSpinner(`🗃`, `All of the template`)
   const list = await fetchRepoList().catch((err)=>{
     console.log(err)
