+++ cli/lib/util/eval.js
