+++ cli/lib/util/logger.js
@@ -1,4 +1,13 @@
 const readline = require('readline')
+const chalk = require('chalk')
+const format = require('util').format
+/**
+ * Prefix.
+ */
+
+const prefix = '   hunyuan-cli'
+const sep = chalk.gray('·')
+
 
 exports.clearConsole = title => {
   if (process.stdout.isTTY) {
@@ -10,4 +19,39 @@ exports.clearConsole = title => {
       console.log(title)
     }
   }
+}
+
+/**
+ * Log a `message` to the console.
+ *
+ * @param {String} message
+ */
+
+exports.log = function (...args) {
+  const msg = format.apply(format, args)
+  console.log(chalk.white(prefix), sep, msg)
+}
+
+/**
+ * Log an error `message` to the console and exit.
+ *
+ * @param {String} message
+ */
+
+exports.fatal = function (...args) {
+  if (args[0] instanceof Error) args[0] = args[0].message.trim()
+  const msg = format.apply(format, args)
+  console.error(chalk.red(prefix), sep, msg)
+  process.exit(1)
+}
+
+/**
+ * Log a success `message` to the console and exit.
+ *
+ * @param {String} message
+ */
+
+exports.success = function (...args) {
+  const msg = format.apply(format, args)
+  console.log(chalk.white(prefix), sep, msg)
 }
