+++ cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@hunyuan/cli",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "description": "Quickly develop command-line tools for front end projects",
   "main": "index.js",
   "bin": {
@@ -18,7 +18,6 @@
     "async": "^3.2.0",
     "axios": "^0.20.0",
     "chalk": "^4.1.0",
-    "cli-prompt": "^0.6.0",
     "commander": "^6.1.0",
     "consolidate": "^0.16.0",
     "download-git-repo": "^3.0.2",
@@ -34,7 +33,6 @@
     "ora": "^5.1.0",
     "read-metadata": "^1.0.0",
     "readline": "^1.3.0",
-    "rimraf": "^3.0.2",
     "user-home": "^2.0.0",
     "validate-npm-package-name": "^3.0.0"
   }
