--- cli/tmp/meta.json
@@ -1,30 +1,0 @@
-{
-  "prompts": {
-    "name": {
-      "type": "string",
-      "required": true,
-      "label": "Project name"
-    },
-    "description": {
-      "type": "string",
-      "required": true,
-      "label": "Project description",
-      "default": "A Vue.js project"
-    },
-    "author": {
-      "type": "string",
-      "label": "Author"
-    },
-    "license": {
-      "type": "string",
-      "label": "License",
-      "default": "MIT"
-    },
-    "sass": {
-       "type": "confirm",
-       "message": "Use sass?",
-       "default": false
-    }
-  },
-  "completeMessage": "{{#inPlace}}To get started:\n\n  npm install\n  npm run dev{{else}}To get started:\n\n  cd {{destDirName}}\n  npm install\n  npm run dev{{/inPlace}}"
-}
