--- cli/tmp/template/.babelrc
@@ -1,6 +1,0 @@
-{
-  "presets": [
-    ["env", { "modules": false }],
-    "stage-3"
-  ]
-}
