--- cli/tmp/template/.gitignore
@@ -1,12 +1,0 @@
-.DS_Store
-node_modules/
-dist/
-npm-debug.log
-yarn-error.log
-
-# Editor directories and files
-.idea
-*.suo
-*.ntvs*
-*.njsproj
-*.sln
