--- cli/tmp/template/src/main.js
@@ -1,7 +1,0 @@
-import Vue from 'vue'
-import App from './App.vue'
-
-new Vue({
-  el: '#app',
-  render: h => h(App)
-})
